backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pypdf2==3.0.1
python-docx==1.1.0
langchain>=0.3.0
langchain-openai>=0.2.0
pydantic>=2.5.3
python-dotenv==1.0.0
